SubmoduleWithBasis.construction: Move here from ModulesWithBasis.Subobjects
