+++ src/sage/categories/modules_with_basis.py
@@ -2425,27 +2425,6 @@ class ModulesWithBasis(CategoryWithAxiom_over_base_ring):
              Category of objects]
         """
 
-        class ParentMethods:
-
-            def construction(self):
-                """
-                Return a pair ``(functor, parent)`` such that ``functor(parent)`` returns ``self``.
-
-                EXAMPLES::
-
-                    sage: M.<a,b> = CombinatorialFreeModule(ZZ, ["a", "b"]); M
-                    Free module generated by {'a', 'b'} over Integer Ring
-                    sage: S = M.submodule([a + 2 * b]); S
-                    Free module generated by {0} over Integer Ring
-                    sage: F, P = S.construction()
-                    sage: F(P) == S
-                    True
-                """
-                from sage.sets.family import Family
-                from sage.categories.pushout import SubspaceFunctor
-                return (SubspaceFunctor(Family(self.basis()).map(self.lift)),
-                        self.ambient())
-
     class CartesianProducts(CartesianProductsCategory):
         """
         The category of modules with basis constructed by Cartesian products
