+++ src/sage/modules/with_basis/subquotient.py
@@ -252,6 +252,25 @@ class SubmoduleWithBasis(CombinatorialFreeModule):
         self.lift_on_basis = self._basis.__getitem__
         self.lift.register_as_coercion()
 
+    def construction(self):
+        """
+        Return a pair ``(functor, parent)`` such that ``functor(parent)`` returns ``self``.
+
+        EXAMPLES::
+
+            sage: M.<a,b> = CombinatorialFreeModule(ZZ, ["a", "b"]); M
+            Free module generated by {'a', 'b'} over Integer Ring
+            sage: S = M.submodule([a + 2 * b]); S
+            Free module generated by {0} over Integer Ring
+            sage: F, P = S.construction()
+            sage: F(P) == S
+            True
+        """
+        from sage.sets.family import Family
+        from sage.categories.pushout import SubspaceFunctor
+        return (SubspaceFunctor(Family(self.basis()).map(self.lift)),
+                self.ambient())
+
     def ambient(self):
         """
         Return the ambient space of ``self``.
